Don't serve the UI with Python. (#371)
This change sets up a dedicated, NGINX container for serving the UI.
NGINX is a great candidate for this, as the files are static and can
be served from disk without executing code on the server.

This also allows us to remove all code from the Python application
related to serving UI files, which makes things simpler.

I tested this locally via the revized instructions for running things
via `docker`, and will deploy an ad-hoc shortly to confirm the changes
in full.

I changed the Python application's root URL to return some information
about the AllenNLP version and the models it's configured to serve.
This information isn't leaking anything, as the repository is already
public, so it felt like a safe change. I can remove this if we feel
otherwise.