Kick off the new paradigm for serving models. (#381)
This includes the code for a small Flask application that serves
predicts using the bidaf model. The application uses some common code
for request logs and configuration that individual applications
shouldn't have to worry about.

I use `docker` to package the resulting application and path munging
such that code can be imported from the `allennlp_demo` module. This
feels a little clunky, and raises some interesting questions around
what happens if an individual model relies on something that doesn't
play nicely with `common`/ -- say, for instance, an old version of
Python.  That said I think in this scenario things will be fine,
as the user would simply modify their `Dockerfile` to not include
the `common` module and instead replace that code with their own,
modified version.

Nothing in the demo packages nor uses this code yet, this is meant
to just get everyone aligned with respect to how I plan on building
out individual endpoints.

An outstanding question is around testing -- in theory I could write
integration tests that very the entire ensemble, but these would likely
take a long time to run given that each endpoint has to download an
archive at startup.