Add K8s config for the new bidaf endpoint. (#383)
This wires up the K8s config required for the bidaf endpoint and
demonstrates how I plan to share code across endpoints.

The `.skiff/bidaf.jsonnet` file is pretty minimal -- to the extent
one might ask if we could omit it all together. I'd like to keep it
as to encourage a clear entry point for configuring each endpoint, so
that if / when customizations are necessary it's clear how to do so.