Add the attack endpoint and abstract out common bits.
This adds code for the `/attack` endpoint. Of note, I couldn't get
the `Hotflip` attacker to work -- but it appears to be broken in
production as well.

As part of this change I split out the common code into
`common.http.ModelEndpoint`, which provides a class that individual
endpoints can either instantiate verbatim or extend.