Add an integration test for the new bidaf endpoint.
Before porting over a new model I thought it'd be nice to get
some tests in place for verying these endpoints.

This introduces a test for the new bidaf endpoint and does so in
a fashion that should make the test code repurposable across endpoints
where possible.

The tests are integration tests and depend on a running server and not
yet wired up to CI -- that'll come later.