Port the bidaf-elmo model to the new serving mechanism. (#389)
This adds code for serving the bidaf-elmo model with the new
mechanism.

The hotflip attack test is currently disabled since it's not
working.

I also removed the `README`s for each individual endpoint as they're
duplicative of the `README` in the `allennlp_demo` package. Having
less means we're more likely to keep them up to date.

Lastly, as part of this I removed quotes from the cloudbuild manifests, 
as they're not required in YAML, and made the model identifier used in 
that file a build substitution variable. This way it's easier to re-use from 
one endpoint to another.