Port the NMN endpoint to the new serving mechanism. (#394)
This migrates the NMN endpoint to the new serving mechanism.

This one is tricky, as currently it's served via a custom image that was
built once. This makes it particularly interesting to port to this
approach, as it validates the value proposition of the effort.

I had to make some changes to common code to accomodate the model.
Specifically:
    - I added support for passing overrides to `load_archive()`.
    - I dynamically import `allennlp_plugins` so that individual
      endpoints can turn this off. In this case we do so because the
      model is sourced from custom code and because it uses an old
      version of AllenNLP that doesn't include the `allennlp_plugins`
      package.

Because of the custom, non `pip` managed dependencies this endpoint
currently only runs (easily) in `Docker`. I've added a `README` that
notes that and provides some commands for running the server and testing
the code.