Refactor permalinks to not store the full response.
This PR adjusts the current permalink mechanism so that the
full response body isn't stored in the database. This means the
UI code is now responsible for reissueing the request to the server.

This has a few benefits:

1. The response data is large. Removing this might significantly reduce
   the database size.
2. It ensures the UI can render the output. If we attempt to rerender
   cached output with a new version of the UI, we can't be sure things
   will work.

As part of this I also removed the `headers` column, as it appears to
have been for diagnostic purposes only and no-one on the team appears to
have used it. This should save some additional space, though I'd expect
it to be negligible.

I tested this with permalinks from prod for two models, coref resolution
and the bidaf-elmo reading comprehension model. Everything worked fine.
I also verified that new permalinks were generated for these two models
as well.

After deploying this I plan to execute the following SQL statements
manually after I've verified that everything is working:

```sql
UPDATE TABLE queries DROP COLUMN headers;
UPDATE TABLE queries DROP COLUMN response_data;
```

Lastly as part of this I made the waiting indicator look a little nicer
(it's a loading spinner instead of text) and fixed the issue where on
larger screens there's a light-grey region on the bottom of the
left-hand and main panes.