Rearrange things a bit.
This moves the old Python application into `old/`, and renames `demo/` to
`ui/`. I split the root level `.skiff/` directory into two pieces, one
chunk now lives in `old/.skiff` and is responsible for deploying the old
application. The other lives in `ui/.skiff` and is responsible for
deploying just the UI bits.

For good measure I also did some cleanup in the `webapp.jsonnet` files
to consolidate to `underscore_based_names` as the use of `camelCase`
and underscores bugged me.