WIP: Break permalinks into a self standing solution.
This is part 1 of the *Great Permalinks Migration*, an epic saga
that tells the tale of a lone traveler, seeking simple, persistent
AllenNLP demo links.

In this PR I move the `/permadata` endpoint into a new, self-standing
service that has no relation to the old code. I tried to change as
little as possible but couldn't help but take a few liberties:

1. I switched the `POST` to a `GET`, which seems more canonical.
2. I replaced the `Permadata` nomenclature with `PermaLink`. This makes
   more sense to me, but let me know if it seems worse, names are hard.

I added a few test cases for the bare-bones.

Still TODO:

1. [ ] Add a `Dockerfile`.
2. [ ] Add documentation and instructions for running things locally.
3. [ ] Make sure the UI actually works (I only ran the tests as a means
       of verifying the code thus far).
4. [ ] Add Kubernetes config and wire up the deploy.
5. [ ] Incorporate Pete's changes, as I imagine this can use the common
       `requirements.txt`.

Part 2, after this lands, involves adding an endpoint for creating
permalinks, at which point I'll change the UI to use and remove all
permalink related code from the old application alltogether.