Part 1 of the Great Permalinks Migration (#412)
This is part 1 of the *Great Permalinks Migration*, an epic saga
that tells the tale of a lone traveler, seeking simple, persistent
AllenNLP demo links.

In this PR I move the `/permadata` endpoint into a new, self-standing
service that has no relation to the old code. I tried to change as
little as possible but couldn't help but take a few liberties:

1. I switched the `POST` to a `GET`, which seems more canonical.
2. I replaced the `Permadata` nomenclature with `PermaLink`. This makes
    more sense to me, but let me know if it seems worse, names are hard.
3. I removed some logic that "health checked" the database before trying to
    query it. I'm pretty sure this was only required because the long-lived db
    connection was being closed after being open for too long. Instead we now
    open a database connection as needed. This is potentially a little slower, but
    a lot less complex, and we've had no issues with the latency associated with
    this in other high traffic demos.

There's also some nomenclature changes to reflect that all endpoints in `api/` are no
longer model specific.

Also, as part of this I got `docker-compose` wired up as a way to set up an end-to-end
local development environment with a single model. I added documentation for this,
please make sure to give that a read. I'm sure there's a few typos in there, as that's
my ™️ after all.

Co-authored-by: Evan Pete Walsh <epwalsh10@gmail.com>