Add new permalinks fields to support the new serving mechanism.
This adds two new fields to the `queries` database table (that which,
rather confusingly, powers the permalinks feature) to support
permalinks for models served by the new mechanism. These fields
are the `task_name` and `model_id`.  The `model_id` will tell the UI
which endpoint to invoke, while the `task_name` is purely for diagnostic
purposes as it'll administrators to reverse engineer a full URL from
the database. Right now that's possible by introspecting the
`request_data`, which isn't terribly easy to do.

I've added the fields such that they're optional. This ensures that
old models can continue creating permalinks as they did before, while
allowing us to add new models that use the new mechanism.

The new fields aren't yet used, but I added a test to make sure they
work. In the next PR I'll transition a the Reading Comprehension demo
to use the new endpoints and make use of this functionality.