Add new permalinks fields to support the new serving mechanism. (#423)
This adds two new fields to the `queries` database table to support
permalinks for models served by the new mechanism. These fields
are `task_name` and `model_id`.  The `model_id` will tell the UI
which endpoint to invoke, while the `task_name` is purely to be used
by administrators to reverse engineer a full URL from the database. 
Right now that's possible by introspecting the `request_data`, but
this isn't terribly easy to do.

The new fields are optional for backwards compatibility. This way the
old models can continue creating permalinks as they did before, and 
we can transition new models to the new mechanism on a case by case
basis.

The new fields aren't yet used, but I added a test to make sure they
work. In the next PR I'll transition a the Reading Comprehension demo
to use the new endpoints and make use of this functionality.  Part of this
will involve migrating the old permalinks to the new format -- I haven't quite
thought through how to do that yet.

I'll need to execute the `ALTER TABLE` command that I added to `schema.sql`
manually as there's no mechanism for executing migrations. This will 
accordingly involve a small window of downtime (the site will still work, 
permalinks will just be broken which the UI and API handle gracefully).