Make it easy to point a local UI at production.
This adds a way to run a local verison of the UI that uses the
production services. This makes it easier to test changes to the
UI that don't require backend updates.

I chose to still run a permalinks service as to keep permalinks
from developer environments from cluttering the production database.