Migrate the RC demo to use the new endpoint. (#427)
This change modifies the RC demo to use the new endpoints. It should
be representative of the work we do demo-by-demo.

I redacted a few identifiers that I changed in the interest of
cleanliness to reduce the amount of work. I also fixed an issue
with routing in the local environment.

After this I plan on adding a local config that proxies all
API requests to production for working on the UI only. This
should make it easier to test changes that don't require working
on the API.