Add the info service.
This adds a simple service that lists all endpoints and some
information we know about each.

The endpoint is a little slow, it takes about 3 or 4 seconds to
collect all of the information. I imagine this is because of the
blocking requests we make to each endpoint. We could either
accept the latency or use concurrency to make it faster. I imagine
making the requests concurrently would greatly speed things up
since they're all I/O bound.