Add Skiff Stats.
This adds support for user metrics provided by Skiff stats.

If this ends up providing the metrics your group needs we can at some point turn off Google Analytics. For now I've left them in place, but let me know if no one is using that (in which case I'd propose we remove them, as it's better for our user's privacy and will make the page a little faster).