Re-use the root certificate.
Prior to this change every endpoint got it's own certificate.
This revises things so that the certificate will be resolved using
the hostname of the request -- which will map to the certificate that's
issued by the UI. This way we're less likely to trigger LetsEncrypt's
rate limits.

This fixes https://github.com/allenai/skiff/issues/843.