Generalize model selection a bit. (#640)
This factors things a bit so that model selection is captured
in the `useModels` hook rather than the task code itself. The
motivation being to reduce the amount of boilerplate that exists
in each demo.

To do this I modified `useModels` to return a class I call a
`ModelList`, which is just a wrapper for a list of models with
a reference to the selected one. This way demos can simply refer
to `list.selected`, rather than managing that value themselves.

The function also returns a second parameter, which is a function
that updates the selected model.

I'm still getting used to using React hooks, but this feels pretty
nice. Curious to see what y'all think.