More useModels factoring. (#644)
This PR commits additional factoring the `useModels()` API to encapsulate more functionality. After this change we also capture code related to requesting predictions from the currently selected model.

This took a bit to put together due to the nuance of `React.useReducer()`.  There's a lot of glue here, the ultimate goal of which is to make the API provided by `useModels()` as easy for the end user to use.  While I don't love all the complexity behind the scenes, I do like the resulting API that the user interacts with.  All of the logic related to handling race conditions created by asynchronous behavior are taken care of transparently, so the user has a simple interface for display purposes.

I snuck in a few extra things too, sorry for the scope creep:

1. I added a top-level `<ErrorBoundary />`. When things `throw` in the actual model code things are caught an ok error message is displayed. It's better than the prior treatment -- a blank page.
2. I added a `<ModelSelect />` control to `tugboat`.