WIP: Less is more.
Factor out the large, complex mechanics of the `useModels` mechanism
I came up with previously in favor of composition.

I also used context here so that in the actual demos themselves the
user has to do very little with respect to handling data flow.
I was skeptical of this, but after getting it working I have to
admit it's pretty nice.