Rework how we handle querying for predictions, take two. (#647)
Factor out the large, complex mechanics of the `useModels` mechanism
(that I previously came up with) into something that favors composition.

I also wired up context so that the data-flow / stateful bits are taken care
of for the "end user" (which in this case is the demo developer). I was 
skeptical of this, but after seeing it in practice I'm in favor of us continuing
to see how this plays out. We still might end up finding it a bit unwieldy, in
which case we can unwind things a bit.