WIP: Add a new endpoint for listing tasks; Update to AllenNLP 1.2.2.
This adds a new endpoint for listing the result of `get_tasks()` so
that the information can be used in the UI.

It doesn't work quite yet, because of bug where the `taskcards`
aren't packaged up in the PyPi package for `allennlp_models`.
The AllenNLP team is aware and will release a new version that
fixes things, at which point we'll update to that version.