Use the tasks endpoint to populate examples and task metadata.
The primary purpose of this PR is to connect the `/api/tasks` endpoint
and the data it provides to thew new, tugboat based demos. Specifics:

- The title and description are now driven by this API, via the
  `<TaskTitle />` and `<TaskDescription />` components.
- The `<SelectExample />` control is now available for showing
  a dropdown with list of examples that can be used. The examples
  are those returned by `/api/tasks` for the endpoint in question.

As part of this I adjusted the `<Promised />` APIs to tolerate
no input for remote things that only need to be fetched once
during the lifetime of the application.