Make sure the model info includes the loaded attackers and interpreters.
Right now the `/info` route for each model returns `null` for the
`attackers` and `interpreters` if they're not specified. For
instance:

```
❯ curl -s https://demo.allennlp.org/api/bidaf | jq 'del(.model_card_data)'
{
  "allennlp": "1.3.0",
  "archive_file": "https://storage.googleapis.com/allennlp-public-models/bidaf-model-2020.03.19.tar.gz",
  "attackers": null,
  "id": "bidaf",
  "interpreters": null,
  "overrides": null,
  "predictor_name": "reading_comprehension",
  "pretrained_model_id": "rc-bidaf",
  "use_old_load_method": false
}
```

This is incorrect, as when they're not specified, the default
interpreters and attackers are loaded instead. This PR fixes that
by assigning these fields default values.

```
❯ curl -s http://localhost:8080/api/bidaf/ | jq 'del(.model_card_data)'
{
  "allennlp": "1.3.0",
  "archive_file": "https://storage.googleapis.com/allennlp-public-models/bidaf-model-2020.03.19.tar.gz",
  "attackers": [
    "hotflip",
    "input_reduction"
  ],
  "id": "bidaf",
  "interpreters": [
    "simple_gradient",
    "smooth_gradient",
    "integrated_gradient"
  ],
  "overrides": null,
  "predictor_name": "reading_comprehension",
  "pretrained_model_id": "rc-bidaf",
  "use_old_load_method": false
}
```

I also tested a model that overrides the field by assigning both
fields to an empty list:

```
❯ cat allennlp_demo/transformer_qa/model.json
{
    "id": "transformer-qa",
    "pretrained_model_id": "rc-transformer-qa",
    "attackers": [],
    "interpreters": []
}
```

This change will allow us to use this in the front-end to toggle
attackers on and off on a model basis. This will allow us to
resolve issues like [this one](https://github.com/allenai/allennlp-demo/issues/679)
in a dyanmic way, rather than hard-coding things to be on or off
on a per-demo basis.