Use the model info to toggle interpreters on and off.
This uses the list of `intepreters` in the `ModelInfo` returned
for the current model to toggle the interpreters that are displayed.

After this change when Transformer QA is selected, the option to
interpret the results no longer displays (which avoids part of this
bug: https://github.com/allenai/allennlp-demo/issues/679).

On the other hand when models like BiDAF or ELMo-BiDAF are selected,
the option to invoke each fo the interpreters remains (since they're
supported for those models).