Add support for model attacks. (#685)
This wires up model attacks in the new demo architecture. As part of
this effort I discovered that the use of `<Form />` was causing some
interesting side-effects. Specifically, the `useEffect()` calls in
`<Form />` were mutating the data in the `<Interpret />` and
`<Attack />` instances. In retrospect this makes sense, and might
suggest that the `<Form />` name is too generic.

I left the `<Form />` abstraction alone and instead decided to wire
`<Attack />` and `<Interpret />` up as straight XHR calls (via
`<Promised />`. This makes it clear there's no user-input in either
case. The `<Interpret />` interface wasn't changed -- as the two
implementation are functionally equivalent.