Add pretrained model id.
The `/api/model-card` endpoint requires that a `pretrained_model_id` be given, which includes the task id as a prefix. This is the key that the `get_pretrained_models()` method uses in the `Dict` it returns. Without this change the model won't be included in the UI, since we don't have a key to use for looking up it's corresponding model card.