Add a button for getting a shareable link. (#713)
This adds a share button, which ultimately produces a permalink
to the UI.

To do this I refactored `<Output.Section />` to be a bit more
flexible. I made the content that can be added to the title
more generic, so that it could be used for the share button in
this case.