Finish up support for shareable URLs (permalinks).
This is Part 2 of 2 (hopefully) of the permalinks effort. It adds
logic that loads and then proceeds to submit the input associated
with a shared URL.

To get this working I had to do a fairly major rearchitecture of
things. Here's the most significant bits that changed:

**Selecting a new model now updates the URL.**

When the user selects a new model using the dropdown, it now adds
the model identifier to the URL path. For instance, selecting BiDAF
takes the user to `/reading-comprehension/bidaf`.

I did this for two reasons:

1. It means we have to persist less data, which means there's less
to restore when loading a shareable URL. Trying to set both the
form input and the selected input made things tricky, since the
bits that manage them are at different levels in the application's
hierarchy. It's definitely possible, it just gets messy fast.

2. It's nice for the end user. The shareable URLs now include the
model identifier, so there's more context that'll help them remember
what the link points to.

**I ripped out the complicated bits of the `usePromise()` API.**

I did this because I discovered a series of bugs related to the
mechanics of the `usePromise()` mechanism. It turns out all the work
I did to try and eliminate race-conditions ultimately made the
API a little unruly *and* made room for some nasty bugs to creep
in.

So I ripped out the complicated bits, but still retained the
`usePromise()` hook and `<Promised />` component, as I think they
still have value. They do a lot less now, but they still save a lot
of boilerplate and seem a lot easier to use now. Plus those bugs
that were around now appear to be gone. Or I might just be tired.

Also, for the record, @aimichal called this when he first saw
these APIs. "Beware ye codeviking who enter here..." couldn't
stop me from treading down the perilous path of premature
generalization.

**A note about the browser history:**

As implemented the history doesn't totally work, but I think it's
ok as-is.

The shareable URLs work great, and can be navigated to via the
browser's history.

If a user loads a shared URL and then submits something else, I
added code that pushes a new history entry into the stack without
the shareable URL parameters. If the user navigates back to this
URL nothing happens. If there's form input or output that's
being displayed it'll continue to be.

I tried really hard to find a way to fix this, but every route I
went down ended up with a big ole' mess of spaghetti. That's why
there's a datestamp at the end of this branch name, as I started
to have a hard time tracking attempts.

Anyway I think it's the right compromise. Overall this feels pretty
nice, and it's worth noting that navigating using the browser
history on the existing demo doesn't work at all -- so it's definitely
an improvement.