Don't refetch things unnecessarily in the UI.
This modifies the UI so that we fetch all of the metadata we need
up front, rather then fetching it per demo. This makes it really
fast to switch from one demo to another, as there's no additional
requests to make. The trade off of course is that there's more going
across the wire in a single request, but the largest of these payloads
is 52 KB, so that should be fine. We ask the user to download a
bunch of JavaScript that's considerably larger than this.

Another benefit of this route is that the requests can all be made
in parallel. So there's an additional speed improvement via that
mechanism too.

As part of this I cleaned a few things up about the the backend too.
The `/api/tasks/all` endpoint is now just `/api/tasks/`, and
`/api/model-card/` became `/api/model-cards/`. This makes the
OCD side of my brain happy, and fixes #734. I also modified
the `/api/info` endpoint to only include information about
the endpoints that do things with models, as the other endpoints
don't really belong there.