Add Sentry for capturing client-side exceptions.
This adds client-side error capturing powered by
[Sentry.io](https://sentry.io).

I added a query string argument that we can use temporarily
to throw an error to make sure things are working. I'll
remove it after we've verified everything.

I also wired things up so that errors should include the
associated environment and SHA, which should make it easier
to debug things.

Lastly I enabled source maps so that the line-numbers
included in the stack traces that Sentry captures are
useful. I can't test this locally, so I'll need to create
an ad-hoc to do so. There's no risk of doing this since
the source code is already public (and arguably there isn't
even if that's not the case, as it's fairly well known
how to reverse minification and obfuscation these days).

Pass through relevant environment variables at build time.