Fix Sentry in production.
I'm not sure why, but right now errors handled by the
`<ErrorBoundary />` aren't being sent to Sentry. This works
in development, but not in production, which leads me to believe
it has something to do with the fact that I decided to *not*
use `<Sentry.ErrorBoundary />`.

This PR factors things out a bit so that we can use the display
logic we already have in place, but use it within the context
of Sentry's `<ErrorBoundary />` I'll test this out in an ad-hoc
to verify that errors get captured.