Fix Sentry in production. (#755)
I'm not sure why, but right now errors handled by the
`<ErrorBoundary />` aren't being sent to Sentry. This works
in development, but not in production, which leads me to believe
it has something to do with the fact that I decided to *not*
use `<Sentry.ErrorBoundary />`.

This PR factors things out a bit so that we can use the display
logic we already have in place, but use it within the context
of Sentry's `<ErrorBoundary />` I'll test this out in an ad-hoc
to verify that errors get captured.

I also added `Sentry.captureException()` calls wherever we
handle exceptions (as this prevents them from bubbling up
to our top-level error boundaries).  This ensures the errors
we cleanly handle (but still want to know about) are captured.