Add the ability to continue LM completions.
This wires up functionality where the completions provided
in the Language Model demo can be continued by clicking them.

To implement this I added a new `Form` context, that gives
access to the immediate `<Form />` parent, if one exists.
We use this context to capture the `antd.FormInstance`
that's attached to that `<Form />`, which allows us to
do things like change the input and submit the form.