Add script for building and managing local containers.
This adds a new script, `./demo`, that can be used to start (and
build) a local container, see a list of those that are runnign (
and those that aren't), and stop those that are running.

At this point things can be started and stopped, and all share
a network, but aren't yet routable from the host machine. I plan
to add that bit in another PR, as this is already a big chunk
of code.