Add a local reverse proxy to the local demo script.
The primary, functional change in this PR adds a local reverse
proxy that handles forwarding requests to either local services
or the production site as appropriate.

But this change also includes a sizeable refactor of the script
(sorry). When I first took a stab at this I started by simply
adding another layer if branching to the existing functions,
but that got really messy and unsustainable quickly. So I decided
to take the plunge and use an object-oriented API instead. I think
the result is a lot easier to follow, and should make it a bit
easier to add additional containers as the demo evolves.

It's also probably worth noting that the script is getting a bit
long, but I like keeping things self-contained and think it's still
worth the tradeoff. Let me know what you all think.