Force usage information to update. (#844)
Fix #827; Use the `key` attribute to force React to ditch the whole
DOM node and rerender it when it changes. This fixes an issue where
Safari wasn't updating the displayed text.

What's bizarre is Safari *was* updating the HTML that's displayed
in the inspector, just not what's rendered. This leads me to think
it's Safari's fault and that React isn't to blame.

It's also worth noting that I tried lots of things, none of which
worked, which led me here:

- I tried replacing the syntax highlighter we use with an alternative
  one (`prism-react-renderer`).
- I tried replacing the syntax highlighting with a simple `<code />`
  tag.

Also, here's a bug that seems somewhat similar and provides further
evidence that this is some sort of low-level, nuanced issue:
https://bugzilla.mozilla.org/show_bug.cgi?id=1105835#c11