Fetch info model by model. (#849)
This modifies the front-end to query info for each model on a
per-model basis. This is probably a bit slower, but also means
that when running the stack locally the info is provided by
the local version that's running instead of production.

If we find it's too slower for users we can probably find other
ways to speed things up.

I also removed the aggregate `/api/info` route. Nothing uses it
after this change, and it involves all sorts of complicated
machinery, so it felt ok to remove.