Fix hook use. (#853)
In f7ab071742e4f917a2e9bf1f06b461470554ce1a I modified `<TaskDemo />`
to fetch the model info as part of it's initialization. When doing so
I missed the fact that I caused a big ole' error to be emitted by
React about the fact that the hooks used in the function body were
not being invoked at the top level. This fixes that error.