Update `@allenai/tugboat` and friends.
This updates `@allenai/tugboat` and it's peers to the latest
versions (which drop Sentry).

Some quick testing suggests things work, but I'd like to do a
bit more testing before releasing this.

I also updated a webpack setting to avoid a deprecation warning
and a few other things.

Of note, I disabled the a11y linting rules because the demo is
likely to get turned off soon. But if we decide to keep the demo
around for longer we should likely re-evaluate that.