Remove Sentry. (#1225)
We don't pay attention to it, and are thus shutting down the account.

To do this I had to update `@allenai/tugboat`, as Sentry was coupled
tightly to that codebase. This required updating a few other things too.

Of note, I disabled a11y linting in a few places because the demo is
likely to get turned off soon. But if we decide to keep the demo
around for longer we should remove those overrides and fix the
issue sinstead.

https://github.com/allenai/reviz/issues/440